@@ -11,39 +11,37 @@
   "devDependencies": {
     "@babel/preset-typescript": "^7.16.7",
     "@nrwl/cli": "13.9.5",
-    "@nrwl/cypress": "13.9.5",
-    "@nrwl/eslint-plugin-nx": "13.9.5",
-    "@nrwl/jest": "13.9.5",
-    "@nrwl/js": "^13.9.5",
-    "@nrwl/linter": "13.9.5",
-    "@nrwl/nx-cloud": "latest",
-    "@nrwl/react": "^13.9.5",
-    "@nrwl/web": "13.9.5",
-    "@nrwl/workspace": "13.9.5",
-    "@testing-library/react": "12.1.4",
-    "@testing-library/react-hooks": "7.0.2",
-    "@types/jest": "27.0.2",
-    "@types/node": "16.11.7",
-    "@types/react": "17.0.40",
-    "@types/react-dom": "17.0.13",
+    "@nrwl/cypress": "14.4.3",
+    "@nrwl/eslint-plugin-nx": "14.4.3",
+    "@nrwl/jest": "14.4.3",
+    "@nrwl/js": "14.4.3",
+    "@nrwl/linter": "14.4.3",
+    "@nrwl/nx-cloud": "14.2.0",
+    "@nrwl/react": "14.4.3",
+    "@nrwl/web": "14.4.3",
+    "@nrwl/workspace": "14.4.3",
+    "@testing-library/react": "13.3.0",
+    "@types/node": "18.0.0",
+    "@types/react": "18.0.14",
+    "@types/react-dom": "18.0.5",
     "@types/uuid": "^8.3.3",
-    "@typescript-eslint/eslint-plugin": "^5.16.0",
-    "@typescript-eslint/parser": "^5.16.0",
-    "babel-jest": "27.2.3",
+    "@typescript-eslint/eslint-plugin": "5.30.6",
+    "@typescript-eslint/parser": "5.30.6",
+    "babel-jest": "27.5.1",
     "cypress": "^9.1.0",
-    "eslint": "~8.7.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
     "eslint-plugin-cypress": "^2.10.3",
-    "eslint-plugin-import": "2.25.4",
-    "eslint-plugin-jsx-a11y": "6.5.1",
-    "eslint-plugin-react": "7.29.3",
-    "eslint-plugin-react-hooks": "4.3.0",
-    "jest": "27.2.3",
-    "nx": "13.9.5",
-    "prettier": "^2.5.1",
-    "react-test-renderer": "17.0.2",
-    "ts-jest": "27.0.5",
-    "typescript": "~4.5.2"
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsx-a11y": "6.6.0",
+    "eslint-plugin-react": "7.30.1",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "jest": "27.5.1",
+    "nx": "14.4.3",
+    "prettier": "2.7.1",
+    "react-test-renderer": "18.2.0",
+    "ts-jest": "27.1.4",
+    "typescript": "4.7.4"
   },
   "dependencies": {
     "@fluentui/react": "^8.37.5",
@@ -56,8 +54,9 @@
     "node-sass": "^6.0.1",
     "ohm-js": "^16.3.1",
     "prettybnf": "^0.1.3",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
+    "react-is": "18.0.0",
     "react-scripts": "4.0.3",
     "regenerator-runtime": "0.13.7",
     "tslib": "^2.0.0",
